backend/requirements.txt
asgiref==3.8.1
azure-core==1.32.0
azure-storage-blob==12.24.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-storages==1.14.5
djangorestframework==3.15.2
easy-thumbnails==2.10
idna==3.10
isodate==0.7.2
pillow==11.1.0
psycopg2==2.9.10
pycparser==2.22
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
